sorts the faces
The previous computation of the faces of the simplices  was done by accumulation them in a set. This had the consequence that the natural ordering of the nodes by their name (when they are named by integers) was not respected when computing the Laplacians. Storing the faces in a sorted list solves the problem.